[bugfix] fixing loading of integrations
It seems I totally missed that integrations have to be classes
and not strings in the parameter. While fixing this I also noticed
that sentry_sdk will have different exception handling for integrations
that are given as parameters.

But it turned out that there is an easier way to prevent the logging
integration from being loaded - so now we monkey patch the sentry_sdk
to remove the logging integration from the defaults.

Also fixed some typos, doc strings, replaced print by logger calls.