[bugfix] fixing loading of integrations
It seems I totally missed that integrations have to be classes
and not strings in the parameter. While fixing this I also noticed
that sentry_sdk will have different exception handling for integrations
that are given as parameters  - so now we monkey patch the sdk
internally to remove the logging integration from the defaults.

Unfortunately this got a bit complicated, as upstream decided to delete
an internal helper after loading the module, so we have to copy the
upstream code verbatim into our module.